Rename kwil-admin peer subcommand to whitelist
