+++ cmd/kwil-admin/cmds/whitelist/cmd.go
@@ -1,14 +1,13 @@
-package peers
+package whitelist
 
 import "github.com/spf13/cobra"
 
 var peersCmd = &cobra.Command{
-	Use:     "peers",
-	Short:   "manages the node's peers",
-	Aliases: []string{"peer"},
+	Use:   "whitelist",
+	Short: "manages the node's whitelist peers",
 }
 
-func PeersCmd() *cobra.Command {
+func WhitelistCmd() *cobra.Command {
 	peersCmd.AddCommand(
 		addCmd(),
 		removeCmd(),
